Remove update from utils
- Moved function and update utility to updater module and not utils. Remains interchangeable as function trigger_update_check remains practically the same, only references need to be updated if moved.
