+++ src/r0tools_simple_toolbox/ext_update/update.py
@@ -1,6 +1,9 @@
 import concurrent.futures
+import json
 import re
 import sys
+import time
+from pathlib import Path
 
 import bpy
 import requests
@@ -24,6 +27,79 @@ def tuple_version_string(version_str: str):
         version_str = match.group(1)
 
     return tuple(int(part) for part in version_str.split("."))
+
+
+def trigger_update_check(*args, **kwargs) -> bool:
+    from ..const import INTERNAL_NAME, REPO_NAME, UPDATE_CHECK_CD
+    from ..ui import r0Tools_PT_SimpleToolbox
+    from ..utils import get_addon_fs_path, get_addon_prefs
+
+    addon_prefs = get_addon_prefs()
+
+    update_check_file: Path = get_addon_fs_path() / INTERNAL_NAME / "check_update"
+    print(f"[INFO] Update File: {str(update_check_file)}")
+
+    KEY_LAST_CHECKED = "last_checked"
+    KEY_CAN_RUN_WHEN = "can_run_when"
+    KEY_UPDATE_AVAILABLE = "update_available"
+
+    update_data = {
+        KEY_LAST_CHECKED: 0,
+        KEY_CAN_RUN_WHEN: 0,
+        KEY_UPDATE_AVAILABLE: False,
+    }
+
+    if addon_prefs.experimental_features and addon_prefs.check_update_startup:
+        print("-------------------------------------------------------")
+        if update_check_file.exists() and update_check_file.is_file():
+            with open(update_check_file, "r") as f:
+                update_data = json.load(f)
+        else:
+            with open(update_check_file, "w") as f:
+                f.write(json.dumps(update_data))
+
+        now = time.time()
+        last_checked = update_data.get(KEY_LAST_CHECKED)
+        can_run_after = update_data.get(KEY_CAN_RUN_WHEN)
+        elapsed_since_check = now - last_checked
+
+        print(f"[INFO] Now: {now:.0f}.")
+        print(f"[INFO] Last checked: {last_checked:.0f}.")
+        print(f"[INFO] Elapsed: {elapsed_since_check:.0f} seconds.")
+
+        if now > can_run_after:
+            has_update = check_extension_update_json(INTERNAL_NAME, REPO_NAME)
+
+            if has_update is None:
+                has_update = False
+
+            update_data[KEY_LAST_CHECKED] = now
+            update_data[KEY_CAN_RUN_WHEN] = now + UPDATE_CHECK_CD
+            update_data[KEY_UPDATE_AVAILABLE] = has_update
+
+            with open(update_check_file, "w") as f:
+                f.write(json.dumps(update_data))
+
+            if has_update:
+                r0Tools_PT_SimpleToolbox._update_callback(has_update)
+
+            return has_update
+        else:
+            remaining_seconds = can_run_after - now
+            print(
+                f"[INFO] Reading from cached file. Can retry in {abs(remaining_seconds):.0f} seconds."
+            )
+            with open(update_check_file, "r") as f:
+                update_data = json.load(f)
+
+            has_update = update_data.get(KEY_UPDATE_AVAILABLE)
+
+            r0Tools_PT_SimpleToolbox._update_callback(has_update)
+        print("-------------------------------------------------------")
+    else:
+        print("Experimental features turned off. No update check.")
+
+    return False
 
 
 def check_extension_update_json(addon_id: str, ext_repo_name: str) -> bool | None:
